requirements.txt
fastapi
uvicorn
requests

python-dotenv
pytest
httpx
sqlalchemy
google-generativeai


langgraph
langchain
langchain-google-genai
langchain-community